refactor: Do not import from lib/index.js
